Make it better

+++ server/db/migrate/20250506000947_add_missing_taggable_index.acts_as_taggable_on_engine.rb
@@ -5,11 +5,15 @@ class AddMissingTaggableIndex < ActiveRecord::Migration[7.1]
   disable_ddl_transaction!
 
   def self.up
-    add_index ActsAsTaggableOn.taggings_table, %i[taggable_id taggable_type context],
-              name: 'taggings_taggable_context_idx', algorithm: :concurrently
+    unless index_exists?(ActsAsTaggableOn.taggings_table, nil, name: 'taggings_taggable_context_idx')
+      add_index ActsAsTaggableOn.taggings_table, %i[taggable_id taggable_type context],
+                name: 'taggings_taggable_context_idx', algorithm: :concurrently
+    end
   end
 
   def self.down
-    remove_index ActsAsTaggableOn.taggings_table, name: 'taggings_taggable_context_idx'
+    if index_exists?(ActsAsTaggableOn.taggings_table, nil, name: 'taggings_taggable_context_idx')
+      remove_index ActsAsTaggableOn.taggings_table, name: 'taggings_taggable_context_idx'
+    end
   end
 end
